Move Config.isLogsMDCTagsInjectionEnabled() to InstrumenterConfig
