+++ dd-java-agent/instrumentation/log4j-2.7/src/main/java/datadog/trace/instrumentation/log4j27/SpanDecoratingContextDataInjector.java
@@ -9,6 +9,7 @@ import static datadog.trace.bootstrap.instrumentation.api.AgentTracer.activeSpan
 import datadog.trace.api.Config;
 import datadog.trace.api.CorrelationIdentifier;
 import datadog.trace.api.DDSpanId;
+import datadog.trace.api.InstrumenterConfig;
 import datadog.trace.bootstrap.instrumentation.api.AgentSpan;
 import datadog.trace.bootstrap.instrumentation.api.Tags;
 import java.util.List;
@@ -32,7 +33,7 @@ public final class SpanDecoratingContextDataInjector implements ContextDataInjec
     // We're at most adding 5 tags
     StringMap newContextData = new SortedArrayStringMap(contextData.size() + 5);
 
-    if (Config.get().isLogsMDCTagsInjectionEnabled()) {
+    if (InstrumenterConfig.get().isLogsMDCTagsInjectionEnabled()) {
       String env = Config.get().getEnv();
       if (null != env && !env.isEmpty()) {
         newContextData.putValue(Tags.DD_ENV, env);
